requirements.txt
python-telegram-bot==20.7
Pillow
numpy
python-dotenv
seam-carving
